Use the correct type for pulse train lengths
Pulse train lengths are defined as uint8_t in the protocol datatype, so
they cannot exceed its range anyway. This unifies the propagation
trough the library to reflect this fact clearly.

NOTE: Users of the PulseTrainCallBack feature need to update their
implementation as the signature of the callback has cahnged. The length
is now proper propagated as uint8_t. See at the changes of
tests/test_parse/test_parse.ino in this commit as an example.

+++ examples/Receive_Raw/Receive_Raw.ino
@@ -12,7 +12,7 @@
 ESPiLight rf(TRANSMITTER_PIN);  //use -1 to disable transmitter
 
 // callback function. It is called on successfully received and parsed rc signal
-void rfRawCallback(const uint16_t* codes, int length) {
+void rfRawCallback(const uint16_t* codes, uint8_t length) {
   // print pulse lengths
   Serial.print("RAW signal: ");
   for(int i=0; i < length; i++) {
